Implemented user input
User can now input into the calculator. The user does not need to
worry about syntax validity as that is handled by the script.
If the user attempts to do an invalid entry, the calculator will not
action. Currently it does not say anything when the user does it.
Future ideas could be to display an error message to make the user aware.
The calculator cannot calculate yet, that will be implemented after this