Implemented a function that processses algbraic expressions
It converts an expression into tokens. Each level of the expression
is contained in an array, subexpressions are contained in their own
array. For example, 2+3-4+(5*5) becomes [2,+,3,-,4,[5,*,5]]. Then
it is reversed which is useful for how it will be solved. Also, created
functions for each operation. It does not fully work with sqrt yet.